#1063 cancel debounce qtabs (#1067)
* #1063 Cancel redraw when component is destroyed

This commit adds a cancel method to the debounce function, allowing to
cancel a previously debounced function to not be executed. The debounced
functions in QTabs are now cancelled just before the component is
destroyed.

* Fixing linter error; Fixed double timeout causing same error

Fixed linter error in debounce.js. Fixed the timeout used to 'let
browser drawing stabilize' by removing it, as the functions are already
debounced and thus delayed. Alternative fix: Save timer id and manually
cancel it in the destroy function.