Update QPage.js
@rstoenescu 
I want to shortly discuss this component.
Today I had to make my app layout float in the middle of the screen when the window is high and wide.
(see screenshot below)

This made me having to override the `min-height` that's encoded on the `q-page` render. I had some trouble doing this! Trying to set `min-height` through style on the element got overridden sometimes by the quasar's style which is here.

This made me think the best way to tackle situations where the dev needs to custom manage `height`, `min-height` and `max-height`.

Instead of adding the possibility for the dev to also set the `height` through a boolean prop, I think it's better to make the developer set those things himself outside of this native component, but then give the option **to disable the `min-height` being set by quasar.**

I think this will give much more freedom.
However, the only problem in that case is that the developer would probably want to have access to the offset:
```js
const offset =
        (this.layout.header.space ? this.layout.header.size : 0) +
        (this.layout.footer.space ? this.layout.footer.size : 0)
```
to be able to keep this in mind when adding custom (min/max)height to the q-page.

So to summarize:

### my suggestion:
"I removed the `height` prop again, set a `minHeight` prop which the dev can set to `false` if he wants to prevent q-page from setting the minHeight. If it's `true` or `undefined` it will just set the minHeight like always.

### My Question:
Is it possible to access this offset somehow in vue components? If we can give the dev access to the offset in a component it's better to write in the docs on how to do that and further not set the props on this component.

PS: Good job on the Vue event today!!


![image](https://user-images.githubusercontent.com/3253920/41044575-8b11457a-69e1-11e8-89b4-1e64fec37791.png)