Update card.ios.styl
If buttons in action area of the card not fitted in one line, they placed on new line and before this button add left margin. This behavior see on first example on Card page on mobile view (button "RESERVE"). This change removes left margin and add right margin on not last button.