Restart work upon reset
When the `reset` method is called, I can only assume that it's because the content of the scrollable area was emptied and that we want to make a fresh new start on loading new content. But if `working` had been set to `false` previously, nothing will happen on the new content (the scripts blocks at line 40). So the `restart` call should make sure that the component is ready to work again.