Add a space before the colon and addition typos
As per French typography, a space is required before all two-signed punctuation (colon, semi-colon, exclamation and question marks). Everything else looks good, except 'Tous' for allRows that can be 'Tous', 'Tout' or 'Toutes' depending on the context. That said, if it indeed refers to the rows, it should be 'Toutes' if it means 'all the rows' or 'Tout' if it means everything. I don't want to make the change for that yet.
I also hope the œ in nœud is not a problem. If it is, oe (o followed by e) is equally as understandable.