Restore "hide-upload-button" option from 0.17
When using the upload component within a larger form, it may be desirable to disallow file uploading directly from the component. 0.17 had this option. Currently 1.0-beta requires the entire header to be overwritten via scoped slot to hide the `upload` button.

This update also hides the `add` button when the `multiple` option is not enabled and a file has already been selected.