added a `thumb-always` option (issue #4002) (#4008)
* added a `thumb-always` option

The default behavior is to hide the slider thumb (the circular marker) whenever a label is present. This can be confusing when the `label-always` option is active, because the thumb is never visible, and it doesn't give the user the expected visual cue that sliding is possible. Even if the user knows, the lack of a thumb makes the visual target (for the mouse, or finger) seem much smaller. Having a `thumb-always` option will keep the thumb visible even with a label, if the developer wants it.

* Update QSlider.json

* Update slider-utils.js