Few word fixes (#4592)
* Removed extra word

* Re-arranged some words

* Not a big deal, can be ignored.

I just felt that the word does not resonate with the day to day meaning of its contextual usage.
Some experts believed that the word "existent" is more common in religious and philosophical writings. 
But again, not too much of a big deal. The consumers of this doc could not care less, I suppose.

* Changed to 'an' article