Improve date checking
Right now your regex expression for date does not actually check for valid dates.  It only checks for valid month, day and year options in a general sense without considering how many days can be in each particular month.  It also does not consider how to handle leap years.

For example:
"2019/02/31" is not a real date, but the old regex would accept this value.

The code revision above should fix these issues.  It is a modified version of a regex expression that I found on StackOverflow (https://stackoverflow.com/questions/15491894/regex-to-validate-date-format-dd-mm-yyyy). 

Note: I used Jason Greanya's regex answer farther down the page (currently has 9 upvotes) as he has a version that is in the correct format (yyyy/MM/dd).  The only changes that I made to his regex is that I required there to be a "/" between year, month and day. (Example: 2019/05/16 = valid; 20190516 = invalid; 2019-05-16 = invalid).

Hope this helps!  As a developer, I really appreciate all that you are doing with quasar.  Congrats on 1.0!