adjust explanation of columns' id 🖍
I had a really hard time understanding the difference between `row-key` which is often set to `name` in the examples here, and the `name` field on the `columns` object.

I know in retrospect it's a really simple difference: columns each have a `name` and rows each have an id or something to be set with `row-key`.

But I believe that the `row-key` has nothing to do with the columns `name` field if I'm not mistaken. So writing "used by row-key" at the columns `name` only further confused me. 😅