fix(app): generator improperly using fs.existsSync
I stumbled across this while preparing the `tauri` merge. The issue is that `fs.existsSync` is there to detect if a FILE exists, not a folder. This is an edge case that shouldn't ever occur, except for cases where the operator makes a mistake or an app-extension / something else is being a bad actor. This PR remediates the improper call and catches this wild edge case, deleting the file that has no reason for existence and then safely creating the folder. 

Note: if .quasar is a folder, nothing changes from current behavior.