Clarify @input event
This caused me some confusion, and now I understand it I'm updating the description to match my current understanding:

Removed reference to panel name, as this doesn't make sense for QStepper or QCarousel.
Clarify that @input event only fires when the component changes the model, not when the model is changed externally.
Note there are other references to 'panel' in other parts of this doc, so perhaps all should be changed to 'panel / slide / step'? Happy to do that if it's useful.