new instructions for windows users
The instructions for windows users were not correct given the state of things in December 2019. In addition to the software listed on this page to install, I had to install the JDK separately and gradle separately. And Cordova has moved on and now requires android-28 SDK to be installed or it wont work. 

I have followed the new instructions in this PR and tested that it worked properly. I pass the cordova requirements check.