feat(extras): Update Ionicons from v4 to v5 (#6425)
The ionicons npm package is missing the woff files. I downloaded their repo created a "repos" folder (npmignore, as well). I am not sure if this is the best way to handle it, but now we can have both v4 and v5 side-by-side. If this repo is updated, be sure to remove the ".git" folder to avoid issues.
If making an announcement, be sure to include this url: https://github.com/ionic-team/ionicons/blob/master/CHANGELOG.md#500-2020-01-15
This includes v5 breaking changes.