Set a default rowsPerPage from rowsPerPageOptions
There's no way to set a default rowsPerPage value without controling/syncing the pagination object.

The innerPagination object in QTable.js sets a default rowsPerPage value as 5 and it doesn't change when using custom rowsPerPageOptions array (For example when using [10, 20, 30, 0] in rowsPerPageOptions, QTable renders 5 rows per page by default). However, it should start with the first setting within the array.

There used to be a rowsPerPage param in QTable, but after the implementation of pagination object, this behaviour is neglected. But there are many cases that pagination object can not be used (in dynamically generated/multiple tables).