docs(fix): Typo in row column wrapping (#7094)
* Fixes typo and changes an example

I tried to come up with a new example, that better demonstrates "column-wrapping". I found the second example confusing for new readers since it also requires knowledge about breakpoints, and to see the effect, you need to resize the browser window. Or you may not even know which rule is applied on the screen you are viewing the docs. So I tried to come up with something more intuitive :) Also fixed a typo.

* Update RowColumnWrapping.vue

Co-authored-by: Razvan Stoenescu <razvan.stoenescu@gmail.com>