first day of the week fix
i made a silly mistake of counting the days from right to left but .split() function separates it from left to right. based on that first day of the week should be index 6/Saturday(شەممە) but I previously set it as index 0 which would make it Friday(هەینی). this is a fix for my mistake.