Update QTime.js
I have had complaints from clients that the QTime is not working logically. The problem is that very often, when the client is selecting the time, it is sharp such as 13:00, 16:00 or 20:00 (or if seconds are used, 13:00:00, 15:30:00, etc). Currently, when the hour is clicked, the minute selection becomes visible, and the minute is null (not selected) by default. At this point the client is practically ready and he/she is willing to click OK button. However, the minute (0) is still required to be selected, which creates unnecessary step for the client. 

Solution: after switching from hour view to minute view, we set the minute to 0 by default (and seconds as well, when entering from minute view to seconds view). This modification is added to this PR.

I don't see any reason why the minute/second should be null after selecting hour/minute.