htmlVariables usage error
In the documentation for the `htmlVariables` option, the example for how to access the variables in index.template.html doesn't work.  The example shows accessing the variables directly be name, but in reality they need to be prefixed by `htmlWebpackPlugin.options.`.  That is, the example in the docs shows: 

```html
<!-- src/index.template.html -->
<%= title %>
<%= some.prop %>
```

But trying this example results in this error: 
```
Template execution failed: ReferenceError: title is not defined
```

This works though:

```html
<!-- src/index.template.html -->
<%= htmlWebpackPlugin.options.title %>
<%= htmlWebpackPlugin.options.some.prop %>
```