fix(app): more refinement for Electron icons
Basically, the "linux-512x512.png" is not needed. The file itself should be renamed to "icon.png", which can be used by all platforms for the Tray icon (the others are needed for the embedded icons). But, we don't want to introduce any breaking changes, so if "icon.png" does not exist, we will copy "linux-512x512.png" to "icon.png". Then, in "electron-main.js" only one image file is needed for the tray icon. Additionally, we can fix "quasar-conf-file.js" to prevent the "can't stat 'icons/icon'" that only happens under linux.