feat(app): new 'electron-handler.js'  for handling 'ipcMain' messages
The 'electron-handler.js' file is being introduced to help developers keep their separation-of-concerns out of the 'electron-main.js' file. It keeps things modular and also introduces a 'useElectronHandler()' method, which keeps it comfortable with those familiar with Vue's composition API. This method is an empty function containing commented example code.