fix(app/icon-genie/docs): more refinement for Electron (#11163)
* fix(app): more refinement for Electron icons

Basically, the "linux-512x512.png" is not needed. The file itself should be renamed to "icon.png", which can be used by all platforms for the Tray icon (the others are needed for the embedded icons). But, we don't want to introduce any breaking changes, so if "icon.png" does not exist, we will copy "linux-512x512.png" to "icon.png". Then, in "electron-main.js" only one image file is needed for the tray icon. Additionally, we can fix "quasar-conf-file.js" to prevent the "can't stat 'icons/icon'" that only happens under linux.

* fix(icon-genie): more refinement for Electron icons

This is to support changes in App in regards to "icon.png"

* feat(app): new 'electron-handler.js'  for handling 'ipcMain' messages

The 'electron-handler.js' file is being introduced to help developers keep their separation-of-concerns out of the 'electron-main.js' file. It keeps things modular and also introduces a 'useElectronHandler()' method, which keeps it comfortable with those familiar with Vue's composition API. This method is an empty function containing commented example code.

* feat(docs): add commented example for adding Node Polyfills

* feat(docs): Give warning about installing into "dependencies" (not "devDependencies")

* feat(docs): updates for icon.png and electron-handler.js

* chore(app) electron template comments update

* Update configuring-electron.md

* Update electron-accessing-files.md

* Update electron-preload-script.md

* Update electron-upgrade-guide.md

* Update electron-upgrade-guide.md

* Update preparation.md

* Update electron.js

* Update electron.js

* Update app-icons-electron.md

* Update electron-main.js

* Delete electron-handler.js

* Update quasar-conf-file.js

* Update quasar-conf-file.js

* Update electron-upgrade-guide.md

* Update preparation.md

Co-authored-by: Razvan Stoenescu <razvan.stoenescu@gmail.com>