fix(app-vite&create-quasar): Utilize tsconfig-preset of app-vite and make fixes (#13689)
* fix: Utilize tsconfig-preset of app-vite
There used to be some problems with 'extends' key, mostly regarding lack of .json extension.
They are fixed, so we should utilize the tsconfig-preset, and do it correctly

* fix(app-vite): Remove 'types' key from tsconfig-preset
This is not needed as vite/client is handled in
create-quasar/templates/app/quasar-v2/ts-vite/BASE/src/shims-vue.d.ts
Also, presence of the 'types' property limits other @types/* from being
loaded/respected in some ways. We can't be sure of this behavioral
changes as it wasn't working for some time, and will be fixed in TS 4.8
So, the best way is to remove it to avoid any future problems too.