fix/feat(vite-plugin): Improve and future-proof the transform checks (#14151)
* fix/feat(vite-plugin): Improve and future-proof the transform checks

* chore(vite-plugin): Add @vitejs/plugin-vue as a peer dep
We are indirectly depending on it through the resource ids. So, if a new
major version of it gets released before a new major version of Vite or
vice-versa, we might not be able to support one or the other.
Making it a peer dep will allows us to selectively support versions.