Update font awesome instructions when using Kits
I have added information on using Font Awesome Pro Kits. In my opinion, this is the way to go when using Font Awesome Pro since it allows you to define a collection of icons that only include icons you use in your project. 

I also included a couple of other tips I wish were written down when I was learning how to use Font Awesome in Quasar. 

If this information was in this doc, it would have saved me at least  3/4 of a day figuring it all out for myself and pestering people on discord.