Fix for issue of Duplicate Workspace in yarn berry
The solution is to create an empty yarn.lock so that src-capacitor is considered to be a separate yarn project and not a workspace of the parent project folder. This is a simple but effective fix to the problem of not being able to add Capacitor mode when using yarn berry.