fix: migrate ecosystem-ci scripts to pnpm
There's no `pnpm workspace` command. Instead, pnpm uses the `--filter`
option to run a command in a specific workspace.

Note there will always be a "ELIFECYCLE Command failed." message at the
end of the vite-plugin test.
But that's from the vite preview server being stopped after a successful
test run. It's not a real error and doesn't affect the test result or
the exit code.